:sparkles: Bugfix and chat error logging (#136)

+++ src/test/kotlin/com/github/pool_party/pull_party_bot/commands/handlers/AbstractBotTest.kt
@@ -228,7 +228,7 @@ internal abstract class AbstractBotTest {
         }
     }
 
-    protected fun verifyMessages(chatId: Long, text: String, exactly: Int = 1, replyTo: Int? = null) {
+    protected fun verifyMessages(chatId: Long, text: String, exactly: Int = 1, replyTo: Long? = null) {
         coVerify(exactly = exactly) {
             bot.sendMessage(chatId, text, any(), any(), any(), any(), replyTo ?: any(), any(), any())
         }
